requirements.txt
bcrypt==4.1.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0

pyotp

